Move internal interceptors to internal.interceptors namespace
