+++ cloudant-client/src/main/java/com/cloudant/client/api/ClientBuilder.java
@@ -25,10 +25,10 @@ import com.cloudant.client.org.lightcouch.CouchDbProperties;
 import com.cloudant.http.HttpConnectionInterceptor;
 import com.cloudant.http.HttpConnectionRequestInterceptor;
 import com.cloudant.http.HttpConnectionResponseInterceptor;
-import com.cloudant.http.interceptors.CookieInterceptor;
-import com.cloudant.http.interceptors.ProxyAuthInterceptor;
-import com.cloudant.http.interceptors.SSLCustomizerInterceptor;
-import com.cloudant.http.interceptors.TimeoutCustomizationInterceptor;
+import com.cloudant.http.internal.interceptors.CookieInterceptor;
+import com.cloudant.http.internal.interceptors.ProxyAuthInterceptor;
+import com.cloudant.http.internal.interceptors.SSLCustomizerInterceptor;
+import com.cloudant.http.internal.interceptors.TimeoutCustomizationInterceptor;
 import com.cloudant.http.internal.interceptors.UserAgentInterceptor;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
