+++ cloudant-http/src/main/java/com/cloudant/http/HttpConnectionRequestInterceptor.java
@@ -17,7 +17,7 @@ package com.cloudant.http;
  * Created by tomblench on 30/03/15.
  */
 
-import com.cloudant.http.interceptors.CookieInterceptor;
+import com.cloudant.http.internal.interceptors.CookieInterceptor;
 
 import java.net.HttpURLConnection;
