+++ cloudant-http/src/main/java/com/cloudant/http/internal/interceptors/HttpConnectionInterceptorException.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2015 IBM Corp. All rights reserved.
+ * Copyright © 2015, 2017 IBM Corp. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
  * except in compliance with the License. You may obtain a copy of the License at
@@ -12,7 +12,7 @@
  * and limitations under the License.
  */
 
-package com.cloudant.http.interceptors;
+package com.cloudant.http.internal.interceptors;
 
 public class HttpConnectionInterceptorException extends RuntimeException {
 
@@ -20,7 +20,7 @@ public class HttpConnectionInterceptorException extends RuntimeException {
     public final String error;
     public final String reason;
 
-    HttpConnectionInterceptorException(Throwable cause) {
+    public HttpConnectionInterceptorException(Throwable cause) {
         this.initCause(cause);
         this.error = null;
         this.reason = null;
